LayoutTests/fast/loader/subframe-removes-itself.html

This tests a subframe that removes itself to make sure a delegate message to indicate the interrupted load fires.
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.dumpFrameLoadCallbacks();
}
</script>
<iframe src="resources/subframe-that-removes-itself.html"></iframe>
